Move lints to their own module
